requirements.txt
absl-py==0.7.0
altgraph==0.16.1
astor==0.7.1
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
Click==7.0
cloudpickle==0.7.0
colorama==0.4.1
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.1.1
decorator==4.3.2
Flask==1.0.2
future==0.17.1
gast==0.2.2
grpcio==1.16.1
h5py==2.8.0
icc-rt==2019.0
imageio==2.4.1
intel-numpy==1.15.1
intel-openmp==2019.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
macholib==1.11
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1
networkx==2.2
numpy==1.15.1
olefile==0.46
opencv-python==4.1.0.25
pandas==0.23.0
parso==0.3.2
patsy==0.5.1
pefile==2018.8.8
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyInstaller==3.4
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
pywin32-ctypes==0.2.0
PyYAML==3.13
pyzmq==17.1.2
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
six==1.12.0
spyder-kernels==0.4.2
statsmodels==0.9.0
tbb==2019.0
tbb4py==2019.0
termcolor==1.1.0
toolz==0.9.0
torch==1.0.1
torchvision==0.2.1
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
wincertstore==0.2
wurlitzer==1.0.2
